Fixing dump file log again
The adress changed again not sure why, as the instruction is always a PUSH that should be pointing to a fixed position in memory(bbcf_base+0x498F04) that contains the minidump comment. This push to said supposedly fixed offset is where we hook the function.

As of now i'm setting it to not care about the second to last digit since it was the only one that changed so far, however it seems like the address is being pushed down so eventually i'll do the better and more robust way of handling it. The ideal and more robust way is during hook initialization scan the memory to find where the string actually is and then hook to the PUSH (0x68) to the found memory.