first step of sanitizing playback data
it was always adding trash to the second byte of each frame's action, because I was assuming it was a char and for some reason was memcpy'ing 2 bytes instead of just one, so the second byte ended up with some random UB trash. It actually isn't 1 byte but 2 bytes. The only use I was able to find for the second byte so far is as a taunt flag, if its set to some values it didn't taunt(such as 0xCC) but for most other did.