states: onhit shouldn't trigger on ukemi actions or pre-actions
added safeguards so that the on hit action wouldn't activate in *Loop* , *Bound*, CmnActBDownCrash and CmnActBDownDown actions. This stops situations where the hitstun is refreshed between states from the states themselves, for example CmdActBDownCrash- > CmdActBDownBound -> CmdActBDownLoop, this is supposed to be the responsability of the wakeup/ukemi action but since there were moments when hitstun ==1 in between them the on hit action would trigger. This is not necessary for the playbacks because on hit actions there function in a different manner.