replay takeover: fix playback not resetting on its own and random replay impeding playback
If the user pressed "Load Replay State" while a previous playback was running the playback wouldn't restart but continue where it was, forcing the user to stop the playback before reloading the takeover state to have consistent playbacks. This is fixed

If the user had its training mode menu's Playback Type set to "Random" the takeover playback would not start. This is fixed.