+ replay upload handshake, settings.ini mutation, visual stuff to select upload on first boot
A lot of work done in the weekend, should be in different commits but forgot.
This will disable replay uploads if p1 or p2 has it disabled, stopping the upload of matches one explicitly does not want.
Adds ways to change settings.ini from within the program.
Adds forced choice on boot on whether should enable replay upload or not.